Add middleware to patch the built-in User model
Because we're making use of the built-in User model for pilots, we need a way
of determining whether a given user is a pilot or not. Providing util.is_pilot
is a good way to start, but it would be more useful if we could query User
instances directly (e.g. particularly in templates). This middleware sets a new
callable on all User instances: is_pilot(self)