Update gpg syntax and add key ID description text
The current version of gpg no longer allows the output flag at the end of the line. This fix resolves that. I also added a bit of context to the 8-digit key ID portions. I thought I had to create the 8-digit key ID but actually I needed to find and copy the one created during the `--full-gen-key` process.